+++ confluence/tree/examples/7-drag_and_drop_a_single_item.tsx
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 // eslint-disable-next-line @atlaskit/ui-styling-standard/use-compiled -- Ignored via go/DSP-18766
 import styled from 'styled-components';
 import Navigation, { AkNavigationItem } from '@atlaskit/navigation';
-import ChevronDownIcon from '@atlaskit/icon/glyph/chevron-down';
-import ChevronRightIcon from '@atlaskit/icon/glyph/chevron-right';
-import DragHandlerIcon from '@atlaskit/icon/glyph/drag-handler';
+import ChevronDownIcon from '@atlaskit/icon/utility/migration/chevron-down';
+import ChevronRightIcon from '@atlaskit/icon/utility/migration/chevron-right';
+import DragHandlerIcon from '@atlaskit/icon/core/migration/drag-handle--drag-handler';
 import Button from '@atlaskit/button/new';
 import Spinner from '@atlaskit/spinner';
 import { token } from '@atlaskit/tokens';
@@ -72,17 +72,19 @@ export default class DragDropWithNestingTree extends Component<void, State> {
 			);
 		}
 		if (item.id === DRAGGABLE_ITEM_ID) {
-			return <DragHandlerIcon label="" size="medium" />;
+			return (
+				<DragHandlerIcon color="currentColor" label="" LEGACY_size="medium" spacing="spacious" />
+			);
 		}
 
 		if (item.children && item.children.length > 0) {
 			return item.isExpanded ? (
 				<Button appearance="subtle" onClick={() => onCollapse(item.id)}>
-					<ChevronDownIcon label="" size="medium" />
+					<ChevronDownIcon color="currentColor" label="" LEGACY_size="medium" />
 				</Button>
 			) : (
 				<Button appearance="subtle" onClick={() => onExpand(item.id)}>
-					<ChevronRightIcon label="" size="medium" />
+					<ChevronRightIcon color="currentColor" label="" LEGACY_size="medium" />
 				</Button>
 			);
 		}