+++ linking-platform/smart-card/examples/6-delayed-loading-indicator.tsx
@@ -85,6 +85,12 @@ class Example extends React.Component<any, ExampleState> {
 								<Card url={this.state.url} appearance="block" />
 								<br />
 								<Card url={this.state.url} appearance="inline" />
+								<br />
+								<Card
+									url={this.state.url}
+									appearance="inline"
+									resolvingPlaceholder={'loading...'}
+								/>
 							</Provider>
 						</div>
 					</GridColumn>