+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/legacy-mobile-macros
 
+## 2.12.0
+
+### Minor Changes
+
+- [#137636](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/pull-requests/137636)
+  [`5bb744ad2d11f`](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/commits/5bb744ad2d11f) -
+  [ux] Enable new icons behind a feature flag.
+
 ## 2.11.37
 
 ### Patch Changes