+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugins
 
+## 4.4.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 4.4.1
 
 ### Patch Changes